Switch to gmavenplus and update groovy
