+++ support/spring/src/test/groovy/org/apache/shiro/spring/config/ShiroBeanConfigurationTest.groovy
@@ -19,6 +19,7 @@
 package org.apache.shiro.spring.config
 
 import org.apache.shiro.event.EventBus
+import org.apache.shiro.spring.testconfig.EventBusConsumersTestConfiguration
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.springframework.beans.factory.annotation.Autowired
