+++ support/spring/src/test/groovy/org/apache/shiro/spring/config/ShiroConfigurationWithOptionalComponentsTest.groovy
@@ -23,6 +23,8 @@ import org.apache.shiro.authz.ModularRealmAuthorizer
 import org.apache.shiro.mgt.DefaultSecurityManager
 import org.apache.shiro.mgt.SecurityManager
 import org.apache.shiro.realm.text.TextConfigurationRealm
+import org.apache.shiro.spring.testconfig.OptionalComponentsTestConfiguration
+import org.apache.shiro.spring.testconfig.RealmTestConfiguration
 import org.apache.shiro.spring.security.interceptor.AuthorizationAttributeSourceAdvisor
 import org.apache.shiro.subject.Subject
 import org.junit.Test
