+++ support/spring/src/test/groovy/org/apache/shiro/spring/web/config/ShiroWebConfigurationTest.groovy
@@ -28,6 +28,8 @@ import org.apache.shiro.spring.config.RealmTestConfiguration
 import org.apache.shiro.spring.config.ShiroAnnotationProcessorConfiguration
 import org.apache.shiro.spring.config.ShiroBeanConfiguration
 import org.apache.shiro.spring.security.interceptor.AuthorizationAttributeSourceAdvisor
+import org.apache.shiro.spring.testconfig.EventBusTestConfiguration
+import org.apache.shiro.spring.testconfig.RealmTestConfiguration
 import org.apache.shiro.web.mgt.CookieRememberMeManager
 import org.apache.shiro.web.mgt.DefaultWebSessionStorageEvaluator
 import org.apache.shiro.web.mgt.WebSecurityManager
