+++ support/spring/src/test/java/org/apache/shiro/spring/testconfig/RealmTestConfiguration.java
@@ -16,46 +16,29 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.shiro.spring.config;
+package org.apache.shiro.spring.testconfig;
 
-
-import org.apache.shiro.event.EventBus;
-import org.apache.shiro.event.EventBusAware;
-import org.apache.shiro.event.Subscribe;
+import org.apache.shiro.realm.Realm;
+import org.apache.shiro.realm.text.TextConfigurationRealm;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.DependsOn;
 
 @Configuration
-public class EventBusConsumersTestConfiguration {
+public class RealmTestConfiguration {
 
     @Bean
-    protected EventBusAwareObject eventBusAwareObject() {
-        return new EventBusAwareObject();
+//    @DependsOn("lifecycleBeanPostProcessor")
+    @SuppressWarnings("Duplicates")
+    Realm getTextConfigurationRealm() {
+
+        TextConfigurationRealm realm = new TextConfigurationRealm();
+        realm.setUserDefinitions("joe.coder=password,user\n" +
+                "jill.coder=password,admin");
+
+        realm.setRoleDefinitions("admin=read,write\n" +
+                "user=read");
+        realm.setCachingEnabled(true);
+        return realm;
     }
-
-    @Bean
-    protected EventSubscriber subscriber(){
-        return new EventSubscriber();
-    }
-
-    public class EventBusAwareObject implements EventBusAware {
-
-        private EventBus eventBus;
-
-        public EventBus getEventBus() {
-            return eventBus;
-        }
-
-        public void setEventBus(EventBus eventBus) {
-            this.eventBus = eventBus;
-        }
-    }
-
-    public class EventSubscriber {
-
-        @Subscribe
-        public void listen(Object object) {}
-
-    }
-
 }
