+++ support/spring/src/test/java/org/apache/shiro/spring/web/testconfig/CacheManagerConfiguration.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.shiro.spring.web.config;
+package org.apache.shiro.spring.web.testconfig;
 
 import org.apache.shiro.cache.CacheManager;
 import org.apache.shiro.cache.MemoryConstrainedCacheManager;
